Move mutable reference (meta)properties to separate module.
(So that they can be imported and played around with in ghci.)
