+++ example/quickcheck-state-machine-example.cabal
@@ -7,6 +7,7 @@ library
   hs-source-dirs:      src
   exposed-modules:     DieHard
                      , MutableReference
+                     , MutableReference.Prop
                      , UnionFind
   build-depends:       base >= 4.7 && < 5
                      , constraints
